# spec/openapi.yaml
openapi: 3.0.3
info:
  title: User Management API
  description: |
    A simple user management API for demonstrating API contract enforcement.
    This spec serves as the source of truth for compliance checking.
  version: 1.0.0
  contact:
    name: API Contract Enforcer

servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: users
    description: User management operations

paths:
  /users:
    get:
      operationId: list_users
      summary: List all users
      description: Retrieve a paginated list of all users
      tags:
        - users
      parameters:
        - name: skip
          in: query
          description: Number of records to skip for pagination
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          description: Maximum number of records to return
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successful response with list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'

    post:
      operationId: create_user
      summary: Create a new user
      description: Create a new user with the provided information
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

  /users/{user_id}:
    get:
      operationId: get_user
      summary: Get user by ID
      description: Retrieve a specific user by their unique identifier
      tags:
        - users
      parameters:
        - name: user_id
          in: path
          description: Unique identifier of the user
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'

    put:
      operationId: update_user
      summary: Update user by ID
      description: Update an existing user's information
      tags:
        - users
      parameters:
        - name: user_id
          in: path
          description: Unique identifier of the user
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

    delete:
      operationId: delete_user
      summary: Delete user by ID
      description: Permanently delete a user from the system
      tags:
        - users
      parameters:
        - name: user_id
          in: path
          description: Unique identifier of the user
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted successfully (no content)
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'

components:
  schemas:
    User:
      type: object
      description: Complete user object with all fields
      required:
        - id
        - email
        - name
        - is_active
        - created_at
      properties:
        id:
          type: integer
          description: Unique identifier for the user
          example: 1
        email:
          type: string
          format: email
          description: User's email address
          example: john@example.com
        name:
          type: string
          description: User's full name
          minLength: 1
          maxLength: 100
          example: John Doe
        is_active:
          type: boolean
          description: Whether the user account is active
          default: true
        created_at:
          type: string
          format: date-time
          description: Timestamp when the user was created
          example: "2024-01-15T10:30:00Z"

    UserCreate:
      type: object
      description: Schema for creating a new user
      required:
        - email
        - name
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: john@example.com
        name:
          type: string
          description: User's full name
          minLength: 1
          maxLength: 100
          example: John Doe
        is_active:
          type: boolean
          description: Whether the user account is active
          default: true

    UserUpdate:
      type: object
      description: Schema for updating an existing user (all fields optional)
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: john.updated@example.com
        name:
          type: string
          description: User's full name
          minLength: 1
          maxLength: 100
          example: John Updated
        is_active:
          type: boolean
          description: Whether the user account is active

    UserList:
      type: object
      description: Paginated list of users
      required:
        - users
        - total
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          description: Total number of users in the system
          example: 42

    HTTPError:
      type: object
      description: Standard HTTP error response
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: User not found

    HTTPValidationError:
      type: object
      description: Validation error response
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'

    ValidationError:
      type: object
      description: Individual validation error
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          description: Location of the error
        msg:
          type: string
          description: Error message
        type:
          type: string
          description: Error type